fix missed from last commit

+++ ww/net/wchecksum.c
@@ -115,7 +115,7 @@ void calcFullPacketChecksum(uint8_t *buf)
     }
 }
 
-uint16_t calcGenericChecksum(const uint8_t *data, size_t len, uint32_t initial)
+uint16_t calcGenericChecksum(const uint8_t *data, uint16_t len, uint32_t initial)
 {
     return checksum(data, len, initial);
 }
